Place squares in the container section
Using an array, multiple grids or boxes were created to fill the
containing <section> element. Currently, all the math to determine the
correct dimensions for the squares are done inside CSS. This will change
later. Event listeners have been added for each square. Currently,
mouseover prints the square's "value" to console. This is used as proof
that the method used to place squares on the grid and the event
listeners for those squares are working. Minimal CSS styling has been
added in this commit.