Add full functionality for changing grid size
In order to achieve the grid resizing functionality, the code that
initializes the grid was placed inside of a function. This function
is called on load with default values. The function is also called
when the user presses the apply button. The values from the slider are
used to generate a new grid after the old grid is "removed." The
removeChild method (as seen in the function removeAllSquares) was
used over innerHTML due to memory leak and security concerns concerning
that method.