Fix outstanding issues related to painting
On previous versions, when painting, the frst div (the initially clicked
div) would not get filled in. This was due to the event handler for
the div firingoff before the pen was activated. This has been fixed by
giving container's event listener a useCapture value of true. An issue
related to accidental dragging has also been fixed by using
e.preventDefault() at the end of the fill function.