Tweak shader behavior add lighten behavior
The way shader works has been tweaked. Instead of editing the alpha
channel for color values, it now increases the individual
rgb color vales by a factor of 255 * 10 (rounded down to 25).
Admittedly, to transform the rgb color string to an array, a regex
pattern found at https://www.30secondsofcode.org/js/s/to-rgb-array was
used.

Using the inverse of the shading algorithm, a dodge or lighten
functionality was added.